backend/requirements.txt
Django==5.2.6
Pillow==11.3.0
djangoresftframework==3.16.1
djangorestframework-simplejwt==5.5.1